+++ graylog2-server/src/main/java/org/graylog/events/notifications/NotificationResourceHandler.java
@@ -17,20 +17,10 @@
 package org.graylog.events.notifications;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
-import org.graylog.events.contentpack.entities.EventProcessorConfigEntity;
-import org.graylog.events.event.EventDto;
-import org.graylog.events.event.EventOriginContext;
 import org.graylog.events.processor.DBEventDefinitionService;
 import org.graylog.events.processor.EventDefinitionDto;
-import org.graylog.events.processor.EventProcessorConfig;
 import org.graylog.scheduler.DBJobDefinitionService;
 import org.graylog.scheduler.JobDefinitionDto;
-import org.graylog2.contentpacks.EntityDescriptorIds;
-import org.graylog2.plugin.Tools;
-import org.graylog2.plugin.rest.ValidationResult;
-import org.graylog2.plugin.streams.Stream;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -51,7 +41,8 @@ public class NotificationResourceHandler {
     @Inject
     public NotificationResourceHandler(DBNotificationService notificationService,
                                        DBJobDefinitionService jobDefinitionService,
-                                       DBEventDefinitionService eventDefinitionService, Map<String, EventNotification.Factory> eventNotificationFactories) {
+                                       DBEventDefinitionService eventDefinitionService,
+                                       Map<String, EventNotification.Factory> eventNotificationFactories) {
         this.notificationService = notificationService;
         this.jobDefinitionService = jobDefinitionService;
         this.eventDefinitionService = eventDefinitionService;
@@ -177,85 +168,20 @@ public class NotificationResourceHandler {
      * @param notificationDto the notification definition to test
      * @param userName the name of the user that triggered the test
      * @throws NotFoundException if the notification definition or the notification factory cannot be found
+     * @throws InternalServerErrorException if the notification definition failed to be executed
      */
-    public void test(NotificationDto notificationDto, String userName) throws NotFoundException {
+    public void test(NotificationDto notificationDto, String userName) throws NotFoundException, InternalServerErrorException {
         final EventNotification.Factory eventNotificationFactory = eventNotificationFactories.get(notificationDto.config().type());
         if (eventNotificationFactory == null) {
             throw new NotFoundException("Couldn't find factory for notification type <" + notificationDto.config().type() + ">");
         }
-        final EventNotificationContext notificationContext = getDummyContext(notificationDto, userName);
+        final EventNotificationContext notificationContext = NotificationTestData.getDummyContext(notificationDto, userName);
         final EventNotification eventNotification = eventNotificationFactory.create();
         try {
             eventNotification.execute(notificationContext);
         } catch (EventNotificationException e) {
             throw new InternalServerErrorException(e.getMessage(), e);
         }
-    }
-
-    private EventNotificationContext getDummyContext(NotificationDto notificationDto, String userName) {
-        final EventDto eventDto = EventDto.builder()
-                .alert(true)
-                .eventDefinitionId("EventDefinitionTestId")
-                .eventDefinitionType("notification-test-v1")
-                .eventTimestamp(Tools.nowUTC())
-                .processingTimestamp(Tools.nowUTC())
-                .id("NotificationTestId")
-                .streams(ImmutableSet.of(Stream.DEFAULT_EVENTS_STREAM_ID))
-                .message("Notification test message triggered from user <" + userName + ">")
-                .source(Stream.DEFAULT_STREAM_ID)
-                .keyTuple(ImmutableList.of("testkey"))
-                .key("testkey")
-                .originContext(EventOriginContext.elasticsearchMessage("testIndex_42", "b5e53442-12bb-4374-90ed-0deadbeefbaz"))
-                .priority(2)
-                .fields(ImmutableMap.of("field1", "value1", "field2", "value2"))
-                .build();
-
-        final EventDefinitionDto eventDefinitionDto = EventDefinitionDto.builder()
-                .alert(true)
-                .id("1234")
-                .title("Event Definition Test Title")
-                .description("Event Definition Test Description")
-                .config(new EventProcessorConfig() {
-                    @Override
-                    public String type() {
-                        return "test-dummy-v1";
-                    }
-                    @Override
-                    public ValidationResult validate() {
-                        return null;
-                    }
-                    @Override
-                    public EventProcessorConfigEntity toContentPackEntity(EntityDescriptorIds entityDescriptorIds) {
-                        return null;
-                    }
-                })
-                .fieldSpec(ImmutableMap.of())
-                .priority(2)
-                .keySpec(ImmutableList.of())
-                .notificationSettings(new EventNotificationSettings() {
-                                          @Override
-                                          public long gracePeriodMs() {
-                                              return 0;
-                                          }
-                                          @Override
-                                          // disable to avoid errors in getBacklogForEvent()
-                                          public long backlogSize() {
-                                              return 0;
-                                          }
-                                          @Override
-                                          public Builder toBuilder() {
-                                              return null;
-                                          }
-                                      }
-
-                ).build();
-
-        return EventNotificationContext.builder()
-                .notificationId("1234")
-                .notificationConfig(notificationDto.config())
-                .event(eventDto)
-                .eventDefinition(eventDefinitionDto)
-                .build();
     }
 
     private EventNotificationExecutionJob.Config getSchedulerConfig(String notificationId) {
